Se agrego un check para parametros para ordernar más carpetas y se corrigio bugs de la carpeta Others, además se agrego una validación para no mover archivos ocultos del sistema

+++ Order.py
@@ -1,7 +1,6 @@
-import os
-import shutil
+import os, shutil, sys, getopt
 
-downloads_path = '/home/carlos/Downloads/'
+downloads_path = None
 images = ('.jpg','.png','.jpeg','.gif')
 video = ('.wmv','.avi','.mp4','.wmv','.mov','.mkv')
 text = ('.doc','.docx','.pdf','.txt')
@@ -27,11 +26,7 @@ def organize(file,ext):
                 shutil.move(downloads_path+file,downloads_path+'/TextFiles/'+file)
 
     else:
-        if os.path.isfile(downloads_path+file):
-            folderCreation('Others')
-            shutil.move(downloads_path+file,downloads_path+'/Others/'+file)
-    if ext =="":
-        if not os.path.isdir(downloads_path+file):
+        if os.path.isfile(downloads_path+file) and not os.path.isdir(downloads_path+file):
             folderCreation('Others')
             shutil.move(downloads_path+file,downloads_path+'/Others/'+file)
 
@@ -42,13 +37,37 @@ def folderCreation(FolderName):
         print(folderName + ' has been created!')
 
 
+def checkArgvs():
+    argv = sys.argv[1:]
+    try:
+        opts, argvs = getopt.getopt(argv,"d:h",['help'])
+    except getopt.GetoptError as Err:
+        print('Error, please check the argument: ',Err)
+        sys.exit(2)
+    for op, ag in opts:
+        global downloads_path
+        if ag == 'dow':
+            downloads_path='/home/carlos/Downloads/'
+            print('Downloads has been established')
+            print(downloads_path)
+        elif ag == 'doc':
+            downloads_path='/home/carlos/'
+            print('Documents has been established')
+        if op in ('-h','--help'):
+            print('Help')
+            
+
+
 def main():
-    for fileX in os.listdir(downloads_path):
-        name, ext = os.path.splitext(fileX)
-        organize(fileX,ext)
-        print(downloads_path+'/'+fileX +' has been moved correctly!')
-    print('-*-*-*-*-*-*-*-*-*-**-*-*-*-*-*-*-*-*')
-    print('All files were organized correctly.!')
+    checkArgvs()
+    if downloads_path != None:
+        for fileX in os.listdir(downloads_path):
+            if not fileX.startswith('.'):
+                name, ext = os.path.splitext(fileX)
+                organize(fileX,ext)
+            #print(downloads_path+'/'+fileX +' has been moved correctly!')
+        print('-*-*-*-*-*-*-*-*-*-**-*-*-*-*-*-*-*-*')
+        print('All files were organized correctly.!')
 
 
 if __name__=='__main__':
